requirements.txt
Django==3.0.6
psycopg2==2.8.5
djangorestframework==3.11.0
